Add jump links
Jump links allow you to specify a specific part of the file to open at.
The syntax is a copy of Go To Anything:
 @ = go to symbol
 : = go to line
 # = go to text

Jump links use both pipelines. First they use the match pipeline to remove
the link from the data. This allows subsequent tests to not have to account
for the link data. Then there is an action in the actions pipeline that moves
the cursor to the extracted location