Remove selecting with the left mouse button
While being able to select with the left mouse button, and then click with the
right mouse button to activate the text is in Acme, it is unnecessarily complex
for this iteration. It adds an extra mouse binding that may get overwritten,
was not documented, and is not exactly obvious.

If people want it, it can be added back later. For now I just want the one binding.